Move Operation to its own file
