+++ modules/internal/src/main/scala/higherkindness/mu/rpc/internal/service/makro/RpcService.scala
@@ -12,36 +12,8 @@ class Helpers[C <: Context](val c: C) {
   val typeAnalysis = new TypeAnalysis[c.type](c)
   import typeAnalysis._
 
-  case class Operation(name: TermName, request: TypeTypology, response: TypeTypology) {
-
-    val isStreaming: Boolean = request.isStreaming || response.isStreaming
-
-    val streamingType: Option[StreamingType] = (request.isStreaming, response.isStreaming) match {
-      case (true, true)  => Some(BidirectionalStreaming)
-      case (true, false) => Some(RequestStreaming)
-      case (false, true) => Some(ResponseStreaming)
-      case _             => None
-    }
-
-    val validStreamingComb: Boolean = (request, response) match {
-      case (_: Fs2StreamTpe, _: MonixObservableTpe) => false
-      case (_: MonixObservableTpe, _: Fs2StreamTpe) => false
-      case _                                        => true
-    }
-
-    require(
-      validStreamingComb,
-      s"RPC service $name has different streaming implementations for request and response"
-    )
-
-    val isMonixObservable: Boolean = List(request, response).collect {
-      case m: MonixObservableTpe => m
-    }.nonEmpty
-
-    val prevalentStreamingTarget: TypeTypology =
-      if (streamingType.contains(ResponseStreaming)) response else request
-
-  }
+  val operationModels = new OperationModels[c.type, typeAnalysis.type](c, typeAnalysis)
+  import operationModels._
 
   class RpcService(serviceDef: ClassDef) {
     val serviceName: TypeName = serviceDef.name
